back/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "iptv"
authors = ["Alexander Kosolapov <kosolapov1976@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.2"
asyncpg = "0.27.0"
bcrypt = "4.0.1"
cffi = "1.15.1"
click = "8.1.3"
colorama = "0.4.6"
cryptography = "40.0.2"
dnspython = "2.3.0"
email-validator = "2.0.0.post2"
fastapi = "0.95.1"
fastapi-users = "11.0.0"
fastapi-users-db-sqlalchemy = "5.0.0"
greenlet = "2.0.2"
h11 = "0.14.0"
httptools = "0.5.0"
idna = "3.4"
makefun = "1.15.1"
passlib = "1.7.4"
pathlib = "1.0.1"
pycparser = "2.21"
pydantic = "1.10.7"
pyjwt = "2.6.0"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
pyyaml = "6.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.12"
starlette = "0.26.1"
typing-extensions = "4.5.0"
uvicorn = "0.22.0"
watchfiles = "0.19.0"
websockets = "11.0.2"
alembic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start= "app.main:start"